+++ src/routes/index.rs
@@ -1,0 +1,9 @@
+use askama_axum::Template;
+
+#[derive(Template)]
+#[template(path = "index.html")]
+pub struct IndexTemplate {}
+
+pub async fn route_handler() -> IndexTemplate {
+    IndexTemplate {}
+}
