+++ src/routes/mod.rs
@@ -1,0 +1,2 @@
+pub mod index;
+pub mod rsvp;
